changed up css for new layout
I've deleted a bunch of code from the index and reduced it only to a few
lines to better see what I'm working with (will add the rest later when
it comes together). Modified a lot of css and deleted some arbitrary
margins and things to compact the look. Also, I changed the result
template, so that when an error occurs, it doesn't sound like it was the
users fault :P.