slimmed the header down
I slimmed down the header so that it's not overly huge and more elegant
looking. Also, I got rid of the Ruby Off Rails link, as I felt it's
already displayed in three different places on the screen (header, About
modal, and footer). Also, I'm committing an odd file for my own
reference. I'll make sure to delete it further down the line!