Created an ugly control panel
Will fix up the CSS for the control panel, it looks hideous. Moved the
results in the info-box on the right and got rid of the footer (and
placed the details in the About modal). Speaking of modals, they are
broken. Will fix those.

Lots of CSS is being cleaned up with this new Foundation. Oh, and got
the grid system to work with me finally... stupid *-* classes...