"control panel" looking much better
Above the right info panel is a small control panel, where we can put
things we deem relevant. It's looks relatively simple.

Trying to figure out a couple things: why that annoying single pixel is there in
between the two panes and also, what to do about the left part.